requirements.txt
Django>=4.2
djangorestframework
Pillow
psycopg2-binary